@@ -4,10 +4,10 @@
   "private": true,
   "homepage": "https://www.dvrpc.org/TIP/NJ/",
   "scripts": {
-    "start": "inferno-scripts start",
-    "build": "inferno-scripts build && npm version patch",
-    "test": "inferno-scripts test --env=jsdom",
-    "eject": "inferno-scripts eject",
+    "start": "react-scripts start",
+    "build": "react-scripts build && npm version patch",
+    "test": "react-scripts test --env=jsdom",
+    "eject": "react-scripts eject",
     "precommit": "lint-staged"
   },
   "lint-staged": {
@@ -18,25 +18,33 @@
   },
   "dependencies": {
     "@mapbox/polyline": "^1.0.0",
-    "eslint": "^5.5.0",
-    "inferno": "^5.5.0",
-    "inferno-clone-vnode": "^5.5.0",
-    "inferno-compat": "^5.5.0",
-    "inferno-create-class": "^5.5.0",
-    "inferno-create-element": "^5.5.0",
-    "inferno-redux": "^5.5.0",
-    "inferno-router": "^5.5.0",
-    "inferno-scripts": "^6.3.0",
     "mapbox-gl": "^0.48.0",
-    "react-autosuggest": "^9.4.0",
+    "react": "^16.12.0",
+    "react-dom": "^16.12.0",
     "react-print": "^1.3.1",
+    "react-redux": "^7.1.3",
+    "react-router-dom": "^5.1.2",
+    "react-select": "^3.0.8",
     "redux": "^4.0.0",
     "redux-thunk": "^2.2.0"
   },
   "devDependencies": {
+    "eslint": "^6.8.0",
     "husky": "^0.15.0-rc.8",
-    "inferno-devtools": "^5.5.0",
     "lint-staged": "^7.0.0",
-    "prettier": "^1.10.2"
+    "prettier": "^1.10.2",
+    "react-devtools": "^4.4.0",
+    "react-scripts": "^3.3.0"
+  },
+  "browserslist": {
+    "production": [
+      ">0.5% in US",
+      "ie 11"
+    ],
+    "development": [
+      "last 1 chrome version",
+      "last 1 firefox version",
+      "last 1 safari version"
+    ]
   }
 }
