@@ -1,7 +1,7 @@
-import { render } from 'inferno';
-import App from './App';
+import { render } from "react";
+import App from "./App";
 
-it('renders without crashing', () => {
-  const div = document.createElement('div');
+it("renders without crashing", () => {
+  const div = document.createElement("div");
   render(<App />, div);
 });
