@@ -1,5 +1,5 @@
-import { Component } from "inferno";
-import { withRouter } from "inferno-router";
+import React, { Component } from "react";
+import { withRouter } from "react-router-dom";
 import "./AppContainer.css";
 
 import Navbar from "../navbar/Navbar.js";
@@ -13,7 +13,7 @@ class AppContainer extends Component {
     const background =
       "background: linear-gradient( to right, white 35%, #E8E8E8 45%, #666)";
     return (
-      <div class="overflow-wrap">
+      <div className="overflow-wrap">
         <Navbar backgroundGradient={background} />
         <div className="appContainer">
           <MapComponentWithRouter />
