@@ -1,11 +1,12 @@
-import { render } from "inferno";
-import { Provider } from "inferno-redux";
+import React from "react";
+import { render } from "react-dom";
+import { Provider } from "react-redux";
 
 import App from "./App";
 import "./index.css";
 import store from "./store";
 
-if (process.env.NODE_ENV === "development") require("inferno-devtools");
+if (process.env.NODE_ENV === "development") require("react-devtools");
 
 render(
   <Provider store={store}>
